Move critical css regeneration out of the endpoint to a new class
